+++ test/DelegationIndexer.t.sol
@@ -11,7 +11,7 @@ contract TestDelegationIndex is Test {
     address coldWallet = 0xC5b9C0549136A7eC4a2270f56afbC29002223F51; // Cold Wallet: wallet that contains the ERC721 tokens
     address hotWallet = 0xe371cDd686341baDbE337D21c53fA51Db505e361; // Wallet that is going to be used as hot wallet 
     address ERC721ContractAddress = 0x3768a0c3d522125f828a3E9F5cA225E4F63fFDb8; // Invented
-    uint256 tokenId = 0;
+    uint32 tokenId = 0;
 
     function setUp() public {
         delegationIndexer = new DelegationIndexer();
